requirements.txt
asgiref @ file:///tmp/build/80754af9/asgiref_1625643473416/work
certifi==2021.10.8
charset-normalizer==2.0.7
colorama==0.4.4
defusedxml==0.7.1
Django @ file:///tmp/build/80754af9/django_1625585912945/work
django-crispy-forms==1.13.0
djangorestframework==3.12.4
httpie==2.6.0
idna==3.3
Pillow==8.4.0
psycopg2 @ file:///C:/ci/psycopg2_1612298766762/work
Pygments==2.10.0
PySocks==1.7.1
pytz==2021.3
requests==2.26.0
requests-toolbelt==0.9.1
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1631814937681/work
urllib3==1.26.7
wincertstore==0.2
